+++ molpipeline/experimental/model_selection/__init__.py
@@ -1,0 +1,6 @@
+"""Model selection module."""
+
+from molpipeline.experimental.model_selection.splitter import GroupShuffleSplit
+
+
+__all__ = ["GroupShuffleSplit"]
