--- molpipeline/model_selection/__init__.py
@@ -1,6 +1,0 @@
-"""Model selection module."""
-
-from molpipeline.model_selection.splitter import GroupShuffleSplit
-
-
-__all__ = ["GroupShuffleSplit"]
